refactor: simplify primality test method call
the chosen method to call for the primality test now works by calling
the value of the radio buttons. this makes the code simpler and easier
to expand.

to achieve this, 'ltk::' had to be added before the function names. this
is just because of how ltk works.

to avoid having to add 'ltk::' before the functions, i could add them
automatically when i call the function in 'generate-prime'. information
on how to achive this is here:
https://stackoverflow.com/questions/15710108/merging-symbols-in-common-lisp