refactor: change how a global variable is defined
the difference between 'defparameter' and 'defvar' is that 'defvar' will
only assign the value to the variable, if the variable doesn't yet
exist.

'*r*' is defined only by its initial value, which is 1. after that, it
should only be changed in the 'lcg' function. so it makes more sense to
use 'defvar' for it, since then it will not be redefined, if the
'defvar' function would be loaded again.