refactor function
instead of using '(integerp (/ random-number j))', it's much faster to
do '(equal (mod random-number j) 0)'.

the rest of the 'naive function has also been rewritten, to reduce side
effects, and make it a little simpler. it also turns out it's a little
faster:

testing with the expression

(time (dotimes (i 10000 i)
        (naive (* 60000 i))))

twice, before changing from 'integerp' and '/' to 'equal' and 'mod', the
previous version of the code executed in an average of 20.257s, and the
new version in 19.69s.

testing with the same expression three times, with the change to using
'equal' and 'mod' instead of 'integerp' and '/', the previous function
executed in an average of 2.876s, and the new version in an average of
2.624s.