refactor function
the 'naive' function no longer has side effects and has been simplified.

one down side of this new implementation is that the random-number that
is passed in must be odd for it to work. to fix this, 'evenp' would have
to be used again in 'naive', or 'logior' would have to be used at least
in two places. the way it is currently set up is simpler.