refactor: remove and add comments
some commented out functions have been removed, as well as some
comments.

a comment has been added, to make it more similar to how it is in the
'generate-random-prime-naive' function.

also 'let*' has been changed to 'let', and the indentation has been
corrected to reflect this. 'let*' creates variables in sequence, instead
of parallel, so it can just be changed to 'let' instead.